修改商家无法保存已拒绝订单的问题， 修复数据库模拟中创建订单时订单状态的错误，添加提交订单时获取新创建订单的指针， 添加订单创建时指定接受者，修改订单未接受状态判断条件， 添加订单管理测试，修改测试模块初始化

+++ Factories/DataSource/SimulateDatabase/SimulateDatabase.h
@@ -72,7 +72,7 @@ private:
 	/* simulate order_states
 	 * order_id	id	last_state	date	transaction	state_name
 	 */
-	std::list<std::tuple<unsigned long, int, unsigned long, std::chrono::system_clock::time_point, double, std::string>> m_orderStates;
+	std::list<std::tuple<unsigned long, int, int, std::chrono::system_clock::time_point, double, std::string>> m_orderStates;
 };
